Replace off64_t with uint64_t where it makes sense to do so.
It looks like off64_t was propagated through the code from the return
type of lseek64(), which isn't appropriate in many of the places we're
using it.

+++ src/common/serve.h
@@ -155,9 +155,8 @@ struct mode_readwrite_params {
 	union mysockaddr     connect_to;
 	union mysockaddr     connect_from;
 
-	/* FIXME: these should be uint64_t and uint32_t respectively */
-	off64_t              from;
-	off64_t              len;
+	uint64_t             from;
+	uint32_t             len;
 
 	int                  data_fd;
 	int                  client;
